+++ packages/jaspr/CHANGELOG.md
@@ -1,6 +1,6 @@
 ## Unreleased breaking
 
-- **Breaking**: Renamed `Curve.linear` to `Curve.linear_` and added `Curve.linear()` function.
+- Added `Curve.linearFn()` easing function.
 - Added `Animation`, `Quotes` CSS properties.
 
 - **Breaking**: `Transition`'s `duration` and `delay` are now of type `Duration` instead of `double`.